spider.Web/pyproject.toml
[project]
name = "spider-web"
version = "0.1.0"
description = "Spider Web — Local multi-agent AI platform with Opus orchestration"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "httpx>=0.27",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "redis>=5.2",
    "pydantic-settings>=2.7",
    "typer>=0.15",
    "rich>=13.9",
    "anthropic>=0.40",
    "openai>=1.55",
    "websockets>=14.0",
    "sse-starlette>=2.1",
    "pyyaml>=6.0",
    "huggingface-hub>=0.25",
    "cryptography>=43.0",
    "mcp>=1.0",
    "plaid-python>=25.0",
    "coinbase-advanced-py>=1.0",
    "oci-openai>=1.0",
]

[project.scripts]
hub = "agentic_hub.cli.app:app"
spiders = "agentic_hub.cli.app:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agentic_hub"]
